Add wasm operator
This commit implements a wasm operator that allows image-based
gadgets to implement custom logic through a wasm module.

The wasm program is stored in a new layer on the OCI image,
it's loaded by this operator. This operator also exposes
some functions that can be used by the guest to access
and modify datasources and fields. Check next commit
to get full list of functions exposed to the modules.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>